web_app/requirements.txt
fastapi
python-jose[cryptography]
pydantic[email]
passlib[bcrypt]
pydantic
python-multipart
uvicorn
jinja2
pymongo
motor
python-dotenv
dnspython